refactor faucet in separate component
