+++ package.json
@@ -10,17 +10,17 @@
     "@hookform/resolvers": "3.3.2",
     "@itheum/sdk-mx-data-nft": "2.4.0",
     "@itheum/sdk-mx-enterprise": "0.0.7",
-    "@multiversx/sdk-core": "12.15.0",
-    "@multiversx/sdk-dapp": "2.24.3",
-    "@multiversx/sdk-native-auth-client": "1.0.5",
-    "@multiversx/sdk-network-providers": "2.2.0",
+    "@multiversx/sdk-core": "12.16.0",
+    "@multiversx/sdk-dapp": "2.24.4",
+    "@multiversx/sdk-native-auth-client": "1.0.6",
+    "@multiversx/sdk-network-providers": "2.1.0",
     "@sentry/react": "7.80.0",
     "@tanstack/match-sorter-utils": "8.8.4",
     "@tanstack/react-table": "8.10.7",
     "@testing-library/jest-dom": "6.1.4",
     "@testing-library/react": "14.1.0",
     "@testing-library/user-event": "14.5.1",
-    "axios": "1.6.1",
+    "axios": "1.6.2",
     "dotenv": "16.3.1",
     "framer-motion": "10.16.4",
     "mime": "3.0.0",
